+++ MK/module/HAL/HAL.h
@@ -32,11 +32,11 @@
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  *
  * Description:          *** HAL for Arduino ***
@@ -92,16 +92,16 @@
       inline void protect() {
         cli();
       }
-  
+
       inline void unprotect() {
         SREG = sreg;
       }
-  
+
       inline InterruptProtectedBlock(bool later = false) {
         sreg = SREG;
         if (!later) cli();
       }
-  
+
       inline ~InterruptProtectedBlock() {
         SREG = sreg;
       }