+++ MK/module/motion/endstops.cpp
@@ -162,6 +162,7 @@ void Endstops::report_state() {
         card.closeFile();
         quickStop();
         disable_all_heaters(); // switch off all heaters.
+        disable_all_coolers();
       }
     #endif
   }