+++ MK/module/motion/planner.h
@@ -100,13 +100,24 @@ typedef struct {
   #endif
 
   #if ENABLED(LASERBEAM)
-    unsigned long laser_ttlmodulation;
-  #endif
+    uint8_t laser_mode; // CONTINUOUS, PULSED, RASTER
+    bool laser_status; // LASER_OFF, LASER_ON
+    float laser_ppm; // pulses per millimeter, for pulsed and raster firing modes
+    unsigned long laser_duration; // laser firing duration in microseconds, for pulsed and raster firing modes
+    unsigned long steps_l; // step count between firings of the laser, for pulsed firing mode
+    float laser_intensity; // Laser firing instensity in clock cycles for the PWM timer
+    #if ENABLED(LASER_RASTER)
+      unsigned char laser_raster_data[LASER_MAX_RASTER_LINE];
+      float laser_raster_intensity_factor;
+    #endif
+  #endif 
 
   volatile char busy;
 } block_t;
 
 #define BLOCK_MOD(n) ((n)&(BLOCK_BUFFER_SIZE-1))
+
+#define MAX_EVENTS_COUNT 2147483648 // max for a signed 32 bit number
 
 // Initialize the motion plan subsystem
 void plan_init();