+++ MK/module/temperature/temperature.h
@@ -45,7 +45,7 @@
 
 // public functions
 void tp_init();  //initialize the heating
-void manage_heater(); //it is critical that this is called periodically.
+void manage_temp_controller(); //it is critical that this is called periodically.
 
 #if ENABLED(FILAMENT_SENSOR)
   // For converting raw Filament Width to milimeters
@@ -69,12 +69,20 @@ void manage_heater(); //it is critical that this is called periodically.
 // do not use these routines and variables outside of temperature.cpp
 extern int target_temperature[4];
 extern float current_temperature[4];
+extern int target_temperature_bed;
+extern float current_temperature_bed;
+extern int target_temperature_chamber;
+extern float current_temperature_chamber;
+extern int target_temperature_cooler;
+extern float current_temperature_cooler;
+
 #if ENABLED(SHOW_TEMP_ADC_VALUES)
   extern int current_temperature_raw[4];
   extern int current_temperature_bed_raw;
+  extern int current_temperature_chamber_raw;
+  extern int current_temperature_cooler_raw;
 #endif
-extern int target_temperature_bed;
-extern float current_temperature_bed;
+
 #if ENABLED(TEMP_SENSOR_1_AS_REDUNDANT)
   extern float redundant_temperature;
 #endif
@@ -92,7 +100,15 @@ extern float current_temperature_bed;
   extern float bedKp, bedKi, bedKd;
 #endif
 
-#if ENABLED(PIDTEMP) || ENABLED(PIDTEMPBED)
+#if ENABLED(PIDTEMPCHAMBER)
+  extern float chamberKp, chamberKi, chamberKd;
+#endif
+
+#if ENABLED(PIDTEMPCOOLER)
+  extern float coolerKp, coolerKi, coolerKd;
+#endif
+
+#if ENABLED(PIDTEMP) || ENABLED(PIDTEMPBED) || ENABLED(PIDTEMPCHAMBER) || ENABLED(PIDTEMPCOOLER)
   float scalePID_i(float i);
   float scalePID_d(float d);
   float unscalePID_i(float i);
@@ -114,14 +130,20 @@ extern float current_temperature_bed;
 
 FORCE_INLINE float degHotend(uint8_t hotend) { return current_temperature[HOTEND_ARG]; }
 FORCE_INLINE float degBed() { return current_temperature_bed; }
+FORCE_INLINE float degChamber() { return current_temperature_chamber; }
+FORCE_INLINE float degCooler() { return current_temperature_cooler; }
 
 #if ENABLED(SHOW_TEMP_ADC_VALUES)
   FORCE_INLINE float rawHotendTemp(uint8_t hotend) { return current_temperature_raw[HOTEND_ARG]; }
   FORCE_INLINE float rawBedTemp() { return current_temperature_bed_raw; }
+  FORCE_INLINE float rawChamberTemp() { return current_temperature_chamber_raw; }
+  FORCE_INLINE float rawCoolerTemp() { return current_temperature_cooler_raw; }
 #endif
 
 FORCE_INLINE float degTargetHotend(uint8_t hotend) { return target_temperature[HOTEND_ARG]; }
 FORCE_INLINE float degTargetBed() { return target_temperature_bed; }
+FORCE_INLINE float degTargetChamber() { return target_temperature_chamber; }
+FORCE_INLINE float degTargetCooler() { return target_temperature_cooler; }
 
 #if ENABLED(THERMAL_PROTECTION_HOTENDS)
   void start_watching_heater(int h = 0);
@@ -129,6 +151,14 @@ FORCE_INLINE float degTargetBed() { return target_temperature_bed; }
 
 #if ENABLED(THERMAL_PROTECTION_BED)
   void start_watching_bed();
+#endif
+
+#if ENABLED(THERMAL_PROTECTION_CHAMBER)
+  void start_watching_chamber();
+#endif
+
+#if ENABLED(THERMAL_PROTECTION_COOLER)
+  void start_watching_cooler();
 #endif
 
 FORCE_INLINE void setTargetHotend(const float& celsius, uint8_t hotend) {
@@ -145,11 +175,29 @@ FORCE_INLINE void setTargetBed(const float& celsius) {
   #endif
 }
 
+FORCE_INLINE void setTargetChamber(const float& celsius) {
+  target_temperature_chamber = celsius;
+  #if ENABLED(THERMAL_PROTECTION_CHAMBER)
+    start_watching_chamber();
+  #endif
+}
+
+FORCE_INLINE void setTargetCooler(const float& celsius) {
+  target_temperature_cooler = celsius;
+  #if ENABLED(THERMAL_PROTECTION_COOLER)
+    start_watching_cooler();
+  #endif
+}
+
 FORCE_INLINE bool isHeatingHotend(uint8_t hotend) { return target_temperature[HOTEND_ARG] > current_temperature[HOTEND_ARG]; }
 FORCE_INLINE bool isHeatingBed() { return target_temperature_bed > current_temperature_bed; }
+FORCE_INLINE bool isHeatingChamber() { return target_temperature_chamber > current_temperature_chamber; }
+FORCE_INLINE bool isHeatingCooler() { return target_temperature_cooler > current_temperature_cooler; } 
 
 FORCE_INLINE bool isCoolingHotend(uint8_t hotend) { return target_temperature[HOTEND_ARG] < current_temperature[HOTEND_ARG]; }
 FORCE_INLINE bool isCoolingBed() { return target_temperature_bed < current_temperature_bed; }
+FORCE_INLINE bool isCoolingChamber() { return target_temperature_chamber < current_temperature_chamber; }
+FORCE_INLINE bool isCoolingCooler() { return target_temperature_cooler < current_temperature_cooler; } 
 
 #define HOTEND_ROUTINES(NR) \
   FORCE_INLINE float degHotend##NR() { return degHotend(NR); } \
@@ -175,11 +223,17 @@ HOTEND_ROUTINES(0);
 #endif
 
 int getHeaterPower(int heater);
+int getBedPower();
+int getChamberPower();
+int getCoolerPower();
+unsigned char getPwmCooler(bool soft);
+
 void disable_all_heaters();
+void disable_all_coolers();
 void updatePID();
 
-#if  HAS(PID_HEATING)
-  void PID_autotune(float temp, int hotend, int ncycles, bool set_result = false);
+#if HAS(PID_HEATING) || HAS(PID_COOLING)
+  void PID_autotune(float temp, int temp_controller, int ncycles, bool set_result = false);
 #endif
 
 void setExtruderAutoFanState(int pin, bool state);