+++ Documentation/changelog.md
@@ -1,3 +1,7 @@
+### Version 4.2.83
+* Add Cooler and Hot Chamber
+* Add Laser Beam PWM and raster base64 ONLY ARDUINO MEGA
+
 ### Version 4.2.82
 * Add DONDOLO_DUAL_MOTOR for DONDOLO bowden and dual extruder
 * Add reader TAG width MFRC522